move class around
