+++ nixops/resources/aws_vpn_connection.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from nixops.state.state import StateDict
+from nixops.state import StateDict
 from nixops.diff import Diff, Handler
 import nixops.util
 import nixops.resources
