+++ nixops/resources/aws_vpn_connection_route.py
@@ -7,7 +7,7 @@ import nixops.resources
 from nixops.resources.ec2_common import EC2CommonState
 import nixops.ec2_utils
 from nixops.diff import Diff, Handler
-from nixops.state.state import StateDict
+from nixops.state import StateDict
 
 class AWSVPNConnectionRouteDefinition(nixops.resources.ResourceDefinition):
     """Definition of a VPN connection route"""
