+++ nixops/resources/vpc_egress_only_internet_gateway.py
@@ -3,7 +3,7 @@
 import boto3
 import botocore
 
-from nixops.state.state import StateDict
+from nixops.state import StateDict
 from nixops.diff import Diff, Handler
 import nixops.util
 import nixops.resources
