+++ nixops/resources/vpc_internet_gateway.py
@@ -7,7 +7,7 @@ import os
 import boto3
 import botocore
 
-from nixops.state.state import StateDict
+from nixops.state import StateDict
 from nixops.diff import Diff, Handler
 import nixops.util
 import nixops.resources
