+++ nixops/resources/vpc_network_interface_attachment.py
@@ -12,7 +12,7 @@ import nixops.resources
 from nixops.resources.ec2_common import EC2CommonState
 import nixops.ec2_utils
 from nixops.diff import Diff, Handler
-from nixops.state.state import StateDict
+from nixops.state import StateDict
 
 class VPCNetworkInterfaceAttachmentDefinition(nixops.resources.ResourceDefinition):
     """Definition of a VPC network interface attachment"""
